maintenance now uses url based routing instead of variables
